# Fix performance issue #2996