--- a/src/Models/HydrostaticFreeSurfaceModels/calculate_hydrostatic_free_surface_tendencies.jl
+++ b/src/Models/HydrostaticFreeSurfaceModels/calculate_hydrostatic_free_surface_tendencies.jl
@@ -262,7 +262,7 @@ end
     @inbounds Gc[i, j, k] =  hydrostatic_free_surface_tracer_tendency(i, j, k, grid, args...)
 end
 
-""" Calculate the right-hand-side of the tracer advection-diffusion equation. """
+""" Calculate the right-hand-side of the subgrid scale energy equation. """
 @kernel function calculate_hydrostatic_free_surface_Ge!(Ge, grid, args)
     i, j, k = @index(Global, NTuple)
     @inbounds Ge[i, j, k] =  hydrostatic_turbulent_kinetic_energy_tendency(i, j, k, grid, args...)