On main we are passing the tendency kernel function as an argument to the kernel. 
Apparently, this prevents compilation on the CPU.
Another place where this design was implemented is the vertically implicit solver, where we pass functions to calculate the tridiagonal matrix coefficients.

This PR fixes both problems.

After this PR we should remember that we cannot pass functions as kernel arguments, 
not even as properties of a struct! Instead we can pass `Val(:function_name)` and dispatch on that to call the correct `function_name` (as this PR implements for the vertically implicit solver)

Baroclinic adjustment test (with `Nx = Ny = 128, Nz = 10`)
on main:
```
[ Info: Initializing simulation...
[00.00%] i: 0, t: 0 seconds, wall time: 870.922 ms, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 16.500 minutes
[ Info:     ... simulation initialization complete (957.942 ms)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (19.178 seconds).
[15.28%] i: 20, t: 5.500 hours, wall time: 1.732 minutes, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 18.150 minutes
[32.08%] i: 40, t: 11.550 hours, wall time: 1.376 minutes, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 19.965 minutes
[50.57%] i: 60, t: 18.205 hours, wall time: 1.333 minutes, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[69.09%] i: 80, t: 1.036 days, wall time: 1.219 minutes, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[87.61%] i: 100, t: 1.314 days, wall time: 1.175 minutes, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[ Info: Simulation is stopping after running for 7.623 minutes.
[ Info: Simulation time 1.500 days equals or exceeds stop time 1.500 days.
```

on this PR:
```
[ Info: Initializing simulation...
[00.00%] i: 0, t: 0 seconds, wall time: 9.474 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 16.500 minutes
[ Info:     ... simulation initialization complete (4.467 seconds)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (29.607 seconds).
[15.28%] i: 20, t: 5.500 hours, wall time: 45.943 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 18.150 minutes
[32.08%] i: 40, t: 11.550 hours, wall time: 14.963 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 19.965 minutes
[50.57%] i: 60, t: 18.205 hours, wall time: 15.328 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[69.09%] i: 80, t: 1.036 days, wall time: 15.417 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[87.61%] i: 100, t: 1.314 days, wall time: 14.862 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[ Info: Simulation is stopping after running for 2.050 minutes.
[ Info: Simulation time 1.500 days equals or exceeds stop time 1.500 days.
```

on Oceananigans v0.79.2 (with KA 0.7.2):
```
[ Info: Initializing simulation...
[00.00%] i: 0, t: 0 seconds, wall time: 52.447 ms, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 16.500 minutes
[ Info:     ... simulation initialization complete (53.391 ms)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (748.379 ms).
[15.28%] i: 20, t: 5.500 hours, wall time: 14.167 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 18.150 minutes
[32.08%] i: 40, t: 11.550 hours, wall time: 14.841 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 19.965 minutes
[50.57%] i: 60, t: 18.205 hours, wall time: 14.606 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[69.09%] i: 80, t: 1.036 days, wall time: 14.924 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[87.61%] i: 100, t: 1.314 days, wall time: 13.987 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 20 minutes
[ Info: Simulation is stopping after running for 1.358 minutes.
[ Info: Simulation time 1.500 days equals or exceeds stop time 1.500 days.
```