@navidcy 2023-03-27
I made a "clean" version of the BCI example:

<details>
  <summary> bench_BCI.jl </summary>
  
  ```julia
  using Oceananigans
  using Oceananigans.Units

  Lx = 1000kilometers
  Ly = 1000kilometers
  Lz = 1kilometers

  Nx = 64
  Ny = 64
  Nz = 40

  grid = RectilinearGrid(size = (Nx, Ny, Nz),
                         x = (0, Lx),
                         y = (-Ly/2, Ly/2),
                         z = (-Lz, 0),
                         topology = (Periodic, Bounded, Bounded))

  model = HydrostaticFreeSurfaceModel(; grid,
                                      coriolis = BetaPlane(latitude = -45),
                                      buoyancy = BuoyancyTracer(),
                                      tracers = :b,
                                      momentum_advection = WENO(),
                                      tracer_advection = WENO())

  ramp(y, Δy) = min(max(0, y/Δy + 1/2), 1)

  N² = 4e-6 # [s⁻²] buoyancy frequency / stratification
  M² = 8e-8 # [s⁻²] horizontal buoyancy gradient

  Δy = 50kilometers # width of the region of the front
  Δb = Δy * M²      # buoyancy jump associated with the front
  ϵb = 1e-2 * Δb    # noise amplitude

  bᵢ(x, y, z) = N² * z + Δb * ramp(y, Δy) + ϵb * randn()

  set!(model, b=bᵢ)

  Δt₀ = 5minutes
  stop_time = 40days

  simulation = Simulation(model, Δt=Δt₀, stop_time=stop_time)
 
  wizard = TimeStepWizard(cfl=0.2, max_change=1.1, max_Δt=20minutes)

  simulation.callbacks[:wizard] = Callback(wizard, IterationInterval(20))

  using Printf

  wall_clock = [time_ns()]

  function print_progress(sim)
      @printf("[%05.2f%%] i: %d, t: %s, wall time: %s, max(u): (%6.3e, %6.3e, %6.3e) m/s, next Δt: %s\n",
              100 * (sim.model.clock.time / sim.stop_time),
              sim.model.clock.iteration,
              prettytime(sim.model.clock.time),
              prettytime(1e-9 * (time_ns() - wall_clock[1])),
              maximum(abs, sim.model.velocities.u),
              maximum(abs, sim.model.velocities.v),
              maximum(abs, sim.model.velocities.w),
              prettytime(sim.Δt))

      wall_clock[1] = time_ns()
    
      return nothing
  end

  simulation.callbacks[:print_progress] = Callback(print_progress, IterationInterval(100))

  @info "Running the simulation..."

  run!(simulation)

  @info "Simulation completed in " * prettytime(simulation.run_wall_time)
  ```
</details>



For me:

### on `main`
```Julia
julia> include("/Users/navid/Desktop/bench_BCI.jl")
[ Info: Precompiling Oceananigans [9e8cae18-63c1-5223-a75c-80ca9d6e9a09]
WARNING: using CUDA.instantiate in module CubedSpheres conflicts with an existing identifier.
[ Info: Oceananigans will use 6 threads
[ Info: Running the simulation...
[ Info: Initializing simulation...
[00.00%] i: 0, t: 0 seconds, wall time: 7.837 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 5.500 minutes
[ Info:     ... simulation initialization complete (3.810 seconds)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (18.936 seconds).
[01.17%] i: 100, t: 11.193 hours, wall time: 1.632 minutes, max(u): (6.985e-01, 3.311e-01, 3.422e-03) m/s, next Δt: 8.858 minutes
[03.04%] i: 200, t: 1.217 days, wall time: 1.307 minutes, max(u): (5.972e-01, 3.536e-01, 2.517e-03) m/s, next Δt: 14.266 minutes
[06.04%] i: 300, t: 2.415 days, wall time: 1.299 minutes, max(u): (5.940e-01, 2.495e-01, 1.809e-03) m/s, next Δt: 20 minutes
[09.51%] i: 400, t: 3.804 days, wall time: 1.314 minutes, max(u): (5.097e-01, 2.361e-01, 1.700e-03) m/s, next Δt: 20 minutes
[12.98%] i: 500, t: 5.193 days, wall time: 1.307 minutes, max(u): (4.790e-01, 2.186e-01, 1.723e-03) m/s, next Δt: 20 minutes
```

### on v0.79.2 (with KernelAbstractions v0.7.2, CUDAKernels v0.3.3)
```Julia
julia> include("/Users/navid/Desktop/bench_BCI.jl")
[ Info: Precompiling Oceananigans [9e8cae18-63c1-5223-a75c-80ca9d6e9a09]
[ Info: Oceananigans will use 6 threads
[ Info: Running the simulation...
[ Info: Initializing simulation...
[00.00%] i: 0, t: 0 seconds, wall time: 6.164 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 5.500 minutes
[ Info:     ... simulation initialization complete (1.943 seconds)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (37.897 seconds).
[01.17%] i: 100, t: 11.193 hours, wall time: 1.616 minutes, max(u): (6.986e-01, 3.323e-01, 3.400e-03) m/s, next Δt: 8.858 minutes
[03.04%] i: 200, t: 1.217 days, wall time: 57.398 seconds, max(u): (5.969e-01, 3.537e-01, 2.494e-03) m/s, next Δt: 14.266 minutes
[06.04%] i: 300, t: 2.415 days, wall time: 57.343 seconds, max(u): (5.982e-01, 2.524e-01, 1.836e-03) m/s, next Δt: 20 minutes
[09.51%] i: 400, t: 3.804 days, wall time: 57.110 seconds, max(u): (5.151e-01, 2.545e-01, 1.683e-03) m/s, next Δt: 20 minutes
[12.98%] i: 500, t: 5.193 days, wall time: 57.194 seconds, max(u): (4.737e-01, 2.410e-01, 1.623e-03) m/s, next Δt: 20 minutes
[16.45%] i: 600, t: 6.581 days, wall time: 56.511 seconds, max(u): (4.713e-01, 2.876e-01, 1.829e-03) m/s, next Δt: 20 minutes
```

### on this PR
```Julia
[ Info: Precompiling Oceananigans [9e8cae18-63c1-5223-a75c-80ca9d6e9a09]
WARNING: using CUDA.instantiate in module CubedSpheres conflicts with an existing identifier.
[ Info: Oceananigans will use 6 threads
[ Info: Running the simulation...
[ Info: Initializing simulation...
[00.00%] i: 0, t: 0 seconds, wall time: 7.729 seconds, max(u): (0.000e+00, 0.000e+00, 0.000e+00) m/s, next Δt: 5.500 minutes
[ Info:     ... simulation initialization complete (3.718 seconds)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (18.285 seconds).
[01.17%] i: 100, t: 11.193 hours, wall time: 1.490 minutes, max(u): (6.969e-01, 3.313e-01, 3.361e-03) m/s, next Δt: 8.858 minutes
[03.04%] i: 200, t: 1.217 days, wall time: 1.174 minutes, max(u): (5.931e-01, 3.540e-01, 2.515e-03) m/s, next Δt: 14.266 minutes
[06.04%] i: 300, t: 2.415 days, wall time: 1.170 minutes, max(u): (5.970e-01, 2.574e-01, 1.871e-03) m/s, next Δt: 20 minutes
[09.51%] i: 400, t: 3.804 days, wall time: 1.175 minutes, max(u): (5.186e-01, 2.392e-01, 1.702e-03) m/s, next Δt: 20 minutes
[12.98%] i: 500, t: 5.193 days, wall time: 1.169 minutes, max(u): (4.768e-01, 2.218e-01, 1.660e-03) m/s, next Δt: 20 minutes
```