@navidcy 2023-03-27
My results seem very different from yours @simone-silvestri. I'll check again later in case I did something wrong...