@simone-silvestri 2023-03-27
To see a larger difference you have to use a vertically implicit closure. 
Anyways, I ll also do some more tests